Fix broken reference list numbering in Firefox
Same as Ib6e9de6.

We must reset the build-in "list-item" counter to make this code behave
sane in Firefox. It looks like this is even described in the CSS spec
and it is not Firefox having a bug, but Chrome being "clever" and not
following the spec.

Bug: T229307
Change-Id: I955786e2b68d087c819a962ded3c571946c61f78
(cherry picked from commit d18c0871a8dfe80b20c70c2121bfadbbe99013fa)