Cite: Move margin in column references
In Vector, Timeless, and Monobook ("desktop"), there is a 0.3em margin applied to <ol>. When in
columns, this causes the first column to be offset from any others vertically. This change moves
the margin from the references <ol> (i.e., sets the list margin-top to 0) to the enclosing div
(sets the div margin-top to 0.3em) when the div has the columns class. This approach is currently
employed by some wikis such as en.wp for their {{reflist}} template.

Minerva doesn't seem to have the relevant top margin.

Bug: T249441
Change-Id: I3eeb3fe9fd690cd8f74df5541b3ef5e10eeed123