ve-cite: Use proper method for comparing linear data (attempt 2)
In Ifb0ba3caf8d we removed this reference equality check and replaced
it with a deep object comparison, however due to the fact that
hashes for MWReferenceNodes were simplified in I8c26686d7b to
improve visual diffs, this resulted in a overly simplified comparison
that couldn't distinguish references with the same "name".

Instead do a deep object comparions of the standard node hash object.

Tests added in Ia0bbd40528 assert that this doesn't result in the
regression that caused T296044.

Bug: T296086
Change-Id: I7b37fb54e14bfe28a07f722a2c45fd4e4a2d44f2