Rename language codes sr-ec and sr-el to sr-cyrl and sr-latn
sr-ec and sr-el are not conform to BCP 47. https://tools.ietf.org/html/bcp47
BCP 47 explicit mentions sr-Latn and sr-Cyrl as examples.

This change renames only the LESS files. The language codes 'sr-cyrl'
and 'sr-latn' are already in the getFallbackLanguages() of the
corresponding language codes.

The i18n JSON files
* from sr-ec.json to sr-cyrl.json and
* from sr-el.json to sr-latn.json
have to renamed in Translatewiki.

Bug: T117845
Depends-On: I666fbdea89ccf21aab6ca1849adf22813dec052e
Change-Id: Ifc310ccad6214db47f6b1487a068f704b3afb460