Fix broken selector in MWReferenceNode.less
This issue was introduced in I7a8ec9f. I tried to understand what this
does but can't see an effect. The behavior and appearence seems to be
identical with and without this code. In case it turns out this is dead
code (e.g. not needed any more with our current browser requirements)
it should be removed. Otherwise the selector fixed.

Change-Id: I0bfae951aa4543c528dd7e04c00a0e978f1ce49e