Render reference list on first load in idle callback
This means the reference list is always in sync with the model's
understanding of which references are available to edit.

An exception is left in if there are no references in the model
at all, as will be the case on he.wiki, as all references are
template generated. In this case we continue to use the Parsoid
DOM, so that there is a static rendering of the reference list.

Bug: T336865
Change-Id: Iaf1089c9de532e7749c9cb70a9e697917955dca8